refactor: 구현체가 아닌 AlarmSender 인터페이스 의존하도록 변경
