+++ src/main/java/net/detalk/api/infrastructure/alarm/discord/DiscordConfig.java
@@ -1,4 +1,4 @@
-package net.detalk.api.support;
+package net.detalk.api.infrastructure.alarm.discord;
 
 import lombok.Getter;
 import lombok.Setter;
